training_service/predictor/requirements.txt
backcall==0.2.0
decorator==5.1.0
entrypoints==0.3
et-xmlfile==1.1.0
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.6
jupyter-core==4.8.1
mysql-connector-python==8.0.26
nest-asyncio==1.5.1
numpy==1.19.5
openpyxl==3.0.9
pandas==1.1.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
protobuf==3.18.1
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
six==1.16.0
tornado==6.1
traitlets==4.3.3
wcwidth==0.2.5
xlrd==2.0.1
sklearn
nltk
flask
bs4==0.0.1
requests==2.26.0
kafka_python==2.0.2
pyspark
sqlalchemy==1.4.25
mysqlclient

